Define what_lens as int, not as size_t
This fixes compiler warnings:

dial.c: In function 'horiz_draw':
dial.c:1269: warning: field precision should have type 'int', but argument
3 has type 'size_t'
file.c: In function 'horiz_draw':
file.c:98: warning: field precision should have type 'int', but argument 3
has type 'size_t'

int should be long enough to hold precision.

By Pavel Roskin <proski@gnu.org>

+++ src/dial.c
@@ -1254,7 +1254,7 @@ static const char *const what[] =
 /* Offsets of what[] entries from position_dialing_directory */
 #define DIALOPTS (sizeof(what) / sizeof(*what))
 #define DIAL_WIDTH 8 /* Width of one entry */
-static size_t what_lens[DIALOPTS]; /* Number of bytes for <= 7 characters */
+static int what_lens[DIALOPTS]; /* Number of bytes for <= 7 characters */
 /* Number of ' ' padding entries at left and right, left is >= 1 */
 static int what_padding[DIALOPTS][2];
